manual sync with trunk - pulling in changes where the issues are not bmesh spesific - some merges added lines in multiple times - removed some NULL checks that were only in bmesh - enable cycles by default (was disabled because it used not to work) - make formatting match

+++ source/blender/blenlib/intern/path_util.c
@@ -35,7 +35,6 @@
 #include <ctype.h>
 #include <string.h>
 #include <stdlib.h>
-#include <stddef.h>
 #include <assert.h>
 
 #include "MEM_guardedalloc.h"
@@ -318,7 +317,7 @@ void BLI_uniquename(ListBase *list, void *vlink, const char defname[], char deli
 
 void BLI_cleanup_path(const char *relabase, char *dir)
 {
-	ptrdiff_t a;
+	short a;
 	char *start, *eind;
 	if (relabase) {
 		BLI_path_abs(dir, relabase);
